Reorganizes more tests, adds dirs based on version
Adds three new directories, common, v16 and 17.

Fixes broken import in test file

Adds dirs and moves files for tests

Adds proper object arg to addnode, not string

More refactoring bitcoind RPC, mostly tests

Splits more tests into their own traits. Adds util function
for creating two unconnected nodes. Useful for checking if
P2P RPC calls are working properly.
