+++ rpc/src/main/scala/org/bitcoins/rpc/client/common/Client.scala
@@ -1,4 +1,4 @@
-package org.bitcoins.rpc.client
+package org.bitcoins.rpc.client.common
 
 import java.util.UUID
 
@@ -11,7 +11,7 @@ import org.bitcoins.core.config.NetworkParameters
 import org.bitcoins.core.crypto.ECPrivateKey
 import org.bitcoins.core.util.BitcoinSLogger
 import org.bitcoins.rpc.config.BitcoindInstance
-import org.bitcoins.rpc.serializers.BitcoindJsonWriters._
+import org.bitcoins.rpc.serializers.BitcoindJsonSerializers._
 import org.slf4j.Logger
 import play.api.libs.json._
