+++ rpc/src/main/scala/org/bitcoins/rpc/client/common/NodeRpc.scala
@@ -1,8 +1,7 @@
-package org.bitcoins.rpc.client
+package org.bitcoins.rpc.client.common
 
 import org.bitcoins.core.number.UInt32
 import org.bitcoins.rpc.jsonmodels.GetMemoryInfoResult
-import org.bitcoins.rpc.serializers.BitcoindJsonReaders._
 import org.bitcoins.rpc.serializers.BitcoindJsonSerializers._
 import play.api.libs.json.{JsArray, JsString, Json}
 
@@ -44,6 +43,10 @@ trait NodeRpc extends Client {
 
   def uptime: Future[UInt32] = {
     bitcoindCall[UInt32]("uptime")
+  }
+
+  def ping(): Future[Unit] = {
+    bitcoindCall[Unit]("ping")
   }
 
   def getMemoryInfo: Future[GetMemoryInfoResult] = {
