+++ rpc/src/main/scala/org/bitcoins/rpc/client/v16/V16AccountRpc.scala
@@ -1,10 +1,8 @@
-package org.bitcoins.rpc.client
+package org.bitcoins.rpc.client.v16
 
-import akka.stream.ActorMaterializer
-import org.bitcoins.core.crypto.DoubleSha256Digest
 import org.bitcoins.core.currency.Bitcoins
 import org.bitcoins.core.protocol.BitcoinAddress
-import org.bitcoins.rpc.config.BitcoindInstance
+import org.bitcoins.rpc.client.common.Client
 import org.bitcoins.rpc.jsonmodels.ReceivedAccount
 import org.bitcoins.rpc.serializers.BitcoindJsonReaders._
 import org.bitcoins.rpc.serializers.BitcoindJsonSerializers._
@@ -12,46 +10,7 @@ import play.api.libs.json.{JsBoolean, JsNumber, JsString}
 
 import scala.concurrent.Future
 
-/**
-  * This class is compatible with version 0.16 of Bitcoin Core.
-  */
-class BitcoindV16RpcClient(override protected val instance: BitcoindInstance)(
-    implicit
-    m: ActorMaterializer)
-    extends BitcoindRpcClient(instance) {
-
-  override val version: BitcoindVersion = BitcoindVersion.V16
-
-  def move(
-      fromAccount: String,
-      toAccount: String,
-      amount: Bitcoins,
-      comment: String = ""): Future[Boolean] = {
-    bitcoindCall[Boolean]("move",
-                          List(JsString(fromAccount),
-                               JsString(toAccount),
-                               JsNumber(amount.toBigDecimal),
-                               JsNumber(6),
-                               JsString(comment)))
-  }
-
-  def sendFrom(
-      fromAccount: String,
-      toAddress: BitcoinAddress,
-      amount: Bitcoins,
-      confirmations: Int = 1,
-      comment: String = "",
-      toComment: String = ""): Future[DoubleSha256Digest] = {
-    bitcoindCall[DoubleSha256Digest](
-      "sendfrom",
-      List(JsString(fromAccount),
-           JsString(toAddress.value),
-           JsNumber(amount.toBigDecimal),
-           JsNumber(confirmations),
-           JsString(comment),
-           JsString(toComment))
-    )
-  }
+trait V16AccountRpc extends Client {
 
   def getAccountAddress(account: String): Future[BitcoinAddress] = {
     bitcoindCall[BitcoinAddress]("getaccountaddress", List(JsString(account)))
