+++ rpc/src/test/scala/org/bitcoins/rpc/v16/BitcoindV16RpcClientTest.scala
@@ -1,11 +1,13 @@
-package org.bitcoins.rpc
+package org.bitcoins.rpc.v16
 
 import akka.actor.ActorSystem
 import akka.stream.ActorMaterializer
 import org.bitcoins.core.config.NetworkParameters
 import org.bitcoins.core.currency.Bitcoins
 import org.bitcoins.core.util.BitcoinSLogger
-import org.bitcoins.rpc.client.BitcoindV16RpcClient
+import org.bitcoins.rpc.TestUtil
+import org.bitcoins.rpc.client.common.RpcOpts.AddNodeArgument
+import org.bitcoins.rpc.client.v16.BitcoindV16RpcClient
 import org.scalatest.{AsyncFlatSpec, BeforeAndAfterAll}
 import org.slf4j.Logger
 
@@ -30,7 +32,7 @@ class BitcoindV16RpcClientTest extends AsyncFlatSpec with BeforeAndAfterAll {
     TestUtil.startServers(client, otherClient)
     logger.info("Bitcoin server started")
 
-    client.addNode(otherClient.getDaemon.uri, "add")
+    client.addNode(otherClient.getDaemon.uri, AddNodeArgument.Add)
 
     logger.info("Funding wallet by mining some blocks")
     Await.result(client.generate(200), 3.seconds)
